FAR-85/refactor:MemberController 변경
